fixed evaluate shape

+++ solver.py
@@ -327,8 +327,8 @@ class Solver():
                 for step, batch in enumerate(self.data_util.train_loader):
                     inputs = batch['input_ids'].to(device)
                     mask = batch['attention_mask'].to(device)
-                    topics  = batch['topic'].to(device).long()
-                    sentiments = batch['sentiment'].to(device).long()
+                    topics  = batch['topic'].to(device)
+                    sentiments = batch['sentiment'].to(device)
 
                     # if (inputs.shape[0] < 32): 
                     #     continue
